Add shared data directory utility and integrate with journal
