+++ src/core/data_dir.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <filesystem>
+
+namespace Core {
+std::filesystem::path resolve_data_dir();
+}
+
